Spring Security. It's alive!

+++ springframe-war/src/main/java/net/tsystems/springframe/frontend/component/OrdersBean.java
@@ -8,11 +8,12 @@
 
 package net.tsystems.springframe.frontend.component;
 
+import net.tsystems.springframe.database.OrderEntity;
 import net.tsystems.springframe.services.Utils;
-import net.tsystems.springframe.services.objects.OrderEntitySO;
-import net.tsystems.springframe.services.objects.OrderexecutorEntitySO;
-import net.tsystems.springframe.services.services.OrderService;
-import net.tsystems.springframe.services.services.OrderexecutorService;
+import net.tsystems.springframe.services.objects.*;
+import net.tsystems.springframe.services.services.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
@@ -23,35 +24,53 @@ import javax.faces.component.html.HtmlInputText;
 import javax.faces.component.html.HtmlSelectOneMenu;
 import java.io.Serializable;
 import java.text.ParseException;
-import java.util.Date;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 
 @SessionScoped
 @ManagedBean(name = "ordersBean")
 @Component(value = "ordersBean")
 public class OrdersBean implements Serializable {
+    private static final Logger logger = LoggerFactory.getLogger(OrdersBean.class);
 
     @Autowired
-    @Qualifier("orderService")
+    public void OrdersBean(OrderService orderService,
+                           OrderexecutorService orderExecutorService,
+                           CargoService cargoService,
+                           TruckService truckService,
+                           DriverService driverService,
+                           RoutepointService routepointService
+                           ){
+        this.orderService = orderService;
+        this.orderExecutorService = orderExecutorService;
+        this.cargoService = cargoService;
+        this.truckService = truckService;
+        this.driverService = driverService;
+        this.routepointService = routepointService;
+
+    }
+
+
     private OrderService orderService;
-
-    @Autowired
-    @Qualifier("orderExecutorService")
     private OrderexecutorService orderExecutorService;
+    private CargoService cargoService;
+    private TruckService truckService;
+    private DriverService driverService;
+    private RoutepointService routepointService;
 
     OrderEntitySO order;
     OrderexecutorEntitySO orderexecutor;
+    DriverEntitySO driver;
+    TruckEntitySO truck;
+    CargoEntitySO cargo;
 
     private HtmlInputText newClosed;
     private HtmlInputText newCreateDT;
     private HtmlInputText newCloseDT;
 
-    private static int editOrderId;
+    private static int editOrderExecutorId;
 
-    public static void setEditOrderId(int editOrderId) {
-        OrdersBean.editOrderId = editOrderId;
+    public static void setEditOrderExecutorId(int editOrderId) {
+        OrdersBean.editOrderExecutorId = editOrderId;
     }
 
     public HtmlInputText getNewClosed() {
@@ -91,16 +110,72 @@ public class OrdersBean implements Serializable {
         return result;
     }
 
-    public Set<OrderEntitySO> getOrder() {
+    public List getOrderExecutors(){
+        List result;
+        try{
+            result = orderExecutorService.getAllOrderexecutors();
+        }catch(NullPointerException npe){
+            npe.printStackTrace();
+            result = null;
+        }
+        return result;
+    }
+
+    public List getWorkingOrderExecutors(){
+         List<CustomOrderExecutor> result;
+        try{
+            List<OrderexecutorEntitySO> executors = orderExecutorService.getWorkingOrderexecutors();
+            HashMap<OrderEntitySO, ArrayList<OrderexecutorEntitySO>> splittedMap = new HashMap<>();
+            for(OrderexecutorEntitySO oe : executors)
+            {
+                if (!splittedMap.containsKey(oe.getOrderIdOrder()))
+                    splittedMap.put(oe.getOrderIdOrder(), new ArrayList<>());
+
+                splittedMap.get(oe.getOrderIdOrder()).add(oe);
+            }
+            if (!splittedMap.isEmpty()) {
+                result = new ArrayList<>();
+                for(Map.Entry<OrderEntitySO, ArrayList<OrderexecutorEntitySO>> e : splittedMap.entrySet()) {
+                    if (!e.getValue().isEmpty()) {
+                        CustomOrderExecutor coe = new CustomOrderExecutor();
+                        coe.setOrderIdOrder(e.getKey());
+                        coe.setTruckIdTruck(e.getValue().get(0).getTruckIdTruck());
+                        StringBuilder driversBuilder = new StringBuilder();
+                        for (int i = 0; i < e.getValue().size(); i++)
+                        {
+                            OrderexecutorEntitySO oe = e.getValue().get(i);
+                            driversBuilder.append(oe.getDriverIdDriver().getName());
+                            driversBuilder.append(' ');
+                            driversBuilder.append(oe.getDriverIdDriver().getSurname());
+                            if ((i + 1) < e.getValue().size())
+                                driversBuilder.append(", \n");
+                            coe.getOrderExecutors().add(oe.getIdOrderExecutor());
+                        }
+                        coe.setDrivers(driversBuilder.toString());
+                        result.add(coe);
+                    }
+                }
+                logger.error("ok");
+                return result;
+            }
+        }catch(NullPointerException npe){
+            npe.printStackTrace();
+            result = null;
+        }
+        return null;
+    }
+
+   // public OrderEntitySO
+
+    public OrderEntitySO getOrderExecutor() {
         if (order == null) {
-            int editId = editOrderId;
+            int editId = editOrderExecutorId;
             this.order = orderService.getOrderById(editId);
             this.newClosed.setValue(this.order.getClosed());
-            this.newCreateDT.setValue(this.order.getCreateDt());
-            this.newCloseDT.setValue(this.order.getCloseDt());
+            this.newCreateDT.setValue(this.order.getCreateDT());
+            this.newCloseDT.setValue(this.order.getCloseDT());
         }
-        Set<OrderEntitySO> result = new HashSet<>();
-        result.add(order);
+        OrderEntitySO result = order;
         return result;
     }
 
@@ -110,27 +185,41 @@ public class OrdersBean implements Serializable {
     }
     public String editOrder() {
         clearItems();
-        int editId = this.editOrderId;
+        int editId = this.editOrderExecutorId;
         this.order = orderService.getOrderById(editId);
         return "editOrder?faces-redirect=true";
     }
 
-    public String delete() {
-        int editId = editOrderId;
-        OrderEntitySO order = orderService.getOrderById(editId);
-        orderService.deleteOrder(order);
+    public String delete(List<Integer> orderExecutorIds) {
+        OrderEntitySO order = null;
+        for(Integer i : orderExecutorIds) {
+            int editId = i;
+            OrderexecutorEntitySO orderExecutor = orderExecutorService.getOrderexecutorById(editId);
+            order = orderExecutor.getOrderIdOrder();
+            orderExecutorService.deleteOrderexecutor(orderExecutor);
+        }
+        if(order!=null) {
+            List<RoutepointEntitySO> routepoints = routepointService.getRoutePointsByOrder(order);
+            if(!routepoints.isEmpty()) {
+                for (RoutepointEntitySO rp : routepoints) {
+                    rp.setOrderIdOrder(null);
+                }
+                routepointService.updateRoutePoints(routepoints);
+            }
+            orderService.deleteOrder(order);
+        }
         return "orders?faces-redirect=true";
     }
 
     public String update(String newClosed, String newCreateDt, String newCloseDt) {
         this.order.setClosed(Utils.byteToBool(Byte.valueOf(newClosed)));
         try {
-            this.order.setCreateDt(Utils.stringToDate(newCreateDt));
+            this.order.setCreateDT(Utils.stringToDate(newCreateDt));
         } catch (ParseException e) {
             e.printStackTrace();
         }
         try {
-            this.order.setCloseDt(Utils.stringToDate(newCloseDt));
+            this.order.setCloseDT(Utils.stringToDate(newCloseDt));
         } catch (ParseException e) {
             e.printStackTrace();
         }
@@ -142,12 +231,12 @@ public class OrdersBean implements Serializable {
         OrderEntitySO order = new OrderEntitySO();
         order.setClosed(Utils.byteToBool(Byte.valueOf(newClosed)));
         try {
-            this.order.setCreateDt(Utils.stringToDate(newCreateDt));
+            this.order.setCreateDT(Utils.stringToDate(newCreateDt));
         } catch (ParseException e) {
             e.printStackTrace();
         }
         try {
-            this.order.setCloseDt(Utils.stringToDate(newCloseDt));
+            this.order.setCloseDT(Utils.stringToDate(newCloseDt));
         } catch (ParseException e) {
             e.printStackTrace();
         }
